Remove Interface suffix and improve implementation name
